# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ColdSafe Monitoring API
  version: 1.0.0
  description: >-
    API do protótipo acadêmico ColdSafe. Os valores e limites são demonstrativos;
    este sistema não é certificado para uso médico, sanitário ou regulatório.
servers:
  - url: http://localhost:8000
    description: Ambiente local
...
paths:
...
  /api/v1/environments:
    get:
      tags:
      - Catalog
      summary: List Environments
      operationId: list_environments_api_v1_environments_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response List Environments Api V1 Environments Get
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - Catalog
      summary: Create Environment
      operationId: create_environment_api_v1_environments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedInput'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Environment Api V1 Environments Post
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /api/v1/environments/{environment_id}:
    get:
      tags:
      - Catalog
      summary: Get Environment
      operationId: get_environment_api_v1_environments__environment_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Environment Api V1 Environments  Environment Id  Get
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Catalog
      summary: Update Environment
      operationId: update_environment_api_v1_environments__environment_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Environment Api V1 Environments  Environment Id  Patch
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/environments/{environment_id}/devices:
    post:
      tags:
      - Catalog
      summary: Create Device
      operationId: create_device_api_v1_environments__environment_id__devices_post
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Device Api V1 Environments  Environment Id  Devices Post
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - Catalog
      summary: List Devices
      operationId: list_devices_api_v1_environments__environment_id__devices_get
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Devices Api V1 Environments  Environment Id  Devices Get
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/devices/{device_id}:
    get:
      tags:
      - Catalog
      summary: Get Device
      operationId: get_device_api_v1_devices__device_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Device Api V1 Devices  Device Id  Get
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Catalog
      summary: Update Device
      operationId: update_device_api_v1_devices__device_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Device Api V1 Devices  Device Id  Patch
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Catalog
      summary: Deactivate Device
      operationId: deactivate_device_api_v1_devices__device_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Deactivate Device Api V1 Devices  Device Id  Delete
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/environments/{environment_id}/rules:
    get:
      tags:
      - Catalog
      summary: Get Rules
      operationId: get_rules_api_v1_environments__environment_id__rules_get
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Rules Api V1 Environments  Environment Id  Rules Get
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - Catalog
      summary: Update Rules
      operationId: update_rules_api_v1_environments__environment_id__rules_put
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThresholdInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Rules Api V1 Environments  Environment Id  Rules Put
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...
  schemas:
    DeviceInput:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        mqtt_device_id:
          type: string
          pattern: ^[A-Za-z0-9._-]{1,64}$
          title: Mqtt Device Id
      additionalProperties: false
      type: object
      required:
      - name
      - mqtt_device_id
      title: DeviceInput
...
    NamedInput:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
      additionalProperties: false
      type: object
      required:
      - name
      title: NamedInput
...
    ThresholdInput:
      properties:
        min_c:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,3}|(?=[\d.]{1,6}0*$)\d{0,3}\.\d{0,2}0*$)
          title: Min C
        max_c:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,3}|(?=[\d.]{1,6}0*$)\d{0,3}\.\d{0,2}0*$)
          title: Max C
        attention_margin_c:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,3}|(?=[\d.]{1,6}0*$)\d{0,3}\.\d{0,2}0*$)
          title: Attention Margin C
      additionalProperties: false
      type: object
      required:
      - min_c
      - max_c
      - attention_margin_c
      title: ThresholdInput
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - DEVICE_NOT_FOUND
        - ENVIRONMENT_NOT_FOUND
        - DEVICE_ID_EXISTS
        - UNAUTHORIZED
        - EMAIL_ALREADY_REGISTERED
        - INVALID_CREDENTIALS
        - DEPENDENCY_UNAVAILABLE
        - INTERNAL_ERROR
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
    ErrorMeta:
      type: object
      additionalProperties: false
      required: [request_id]
      properties:
        request_id:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
        meta:
          $ref: "#/components/schemas/ErrorMeta"